add worker service
