+++ mars/scheduler/service.py
@@ -25,47 +25,47 @@ from .kvstore import KVStoreActor
 from ..node_info import NodeInfoActor
 
 
-def start_service(endpoint, pool, app=None):
-    """
-    there are two way to start a scheduler
-    1) if options.kv_store is specified as an etcd address, the endpoint will be written
-    into kv-storage to indicate that this scheduler is one the schedulers,
-    and the etcd is used as a service discover.
-    2) if options.kv_store is not an etcd address, there will be only one scheduler
-    """
-    kv_store = kvstore.get(options.kv_store)
-    if isinstance(kv_store, kvstore.EtcdKVStore):
-        # set etcd as service discover
-        service_discover_addr = options.kv_store
-        schedulers = None
-        kv_store.write('/schedulers/%s' % endpoint, dir=True)
-    else:
-        # single scheduler
-        service_discover_addr = None
-        schedulers = [endpoint]
+class SchedulerService(object):
+    def __init__(self):
+        self._cluster_info_ref = None
+        self._session_manager_ref = None
+        self._resource_ref = None
+        self._kv_store_ref = None
+        self._node_info_ref = None
 
-    # create ClusterInfoActor
-    cluster_info_ref = pool.create_actor(
-        ClusterInfoActor, schedulers, service_discover_addr, uid=ClusterInfoActor.default_name())
-    if app is not None:
-        app._cluster_info_ref = cluster_info_ref
-    # create SessionManagerActor
-    session_manager_ref = pool.create_actor(SessionManagerActor,
-                                            uid=SessionManagerActor.default_name())
-    if app is not None:
-        app._session_manager_ref = session_manager_ref
-    # create ResourceActor
-    resource_ref = pool.create_actor(ResourceActor, uid=ResourceActor.default_name())
-    if app is not None:
-        app._resource_ref = resource_ref
-    # create KVStoreActor
-    kv_store_ref = pool.create_actor(KVStoreActor, uid=KVStoreActor.default_name())
-    if app is not None:
-        app._kv_store_ref = kv_store_ref
-    # create NodeInfoActor
-    node_info_ref = pool.create_actor(NodeInfoActor, uid=NodeInfoActor.default_name())
-    if app is not None:
-        app._node_info_ref = node_info_ref
+    def start(self, endpoint, pool):
+        """
+        there are two way to start a scheduler
+        1) if options.kv_store is specified as an etcd address, the endpoint will be written
+        into kv-storage to indicate that this scheduler is one the schedulers,
+        and the etcd is used as a service discover.
+        2) if options.kv_store is not an etcd address, there will be only one scheduler
+        """
+        kv_store = kvstore.get(options.kv_store)
+        if isinstance(kv_store, kvstore.EtcdKVStore):
+            # set etcd as service discover
+            service_discover_addr = options.kv_store
+            schedulers = None
+            kv_store.write('/schedulers/%s' % endpoint, dir=True)
+        else:
+            # single scheduler
+            service_discover_addr = None
+            schedulers = [endpoint]
 
-    kv_store.write('/schedulers/%s/meta' % endpoint,
-                   json.dumps(resource_ref.get_workers_meta()))
+        # create ClusterInfoActor
+        self._cluster_info_ref = pool.create_actor(
+            ClusterInfoActor, schedulers, service_discover_addr, uid=ClusterInfoActor.default_name())
+        # create SessionManagerActor
+        self._session_manager_ref = pool.create_actor(SessionManagerActor,
+                                                uid=SessionManagerActor.default_name())
+        # create ResourceActor
+        self._resource_ref = pool.create_actor(ResourceActor, uid=ResourceActor.default_name())
+        # create KVStoreActor
+        self._kv_store_ref = pool.create_actor(KVStoreActor, uid=KVStoreActor.default_name())
+        # create NodeInfoActor
+        self._node_info_ref = pool.create_actor(NodeInfoActor, uid=NodeInfoActor.default_name())
+        kv_store.write('/schedulers/%s/meta' % endpoint,
+                       json.dumps(self._resource_ref.get_workers_meta()))
+
+    def stop(self, pool):
+        pool.destroy_actor(self._resource_ref)
