+++ mars/scheduler/tests/test_main.py
@@ -51,7 +51,6 @@ class Test(unittest.TestCase):
             shutil.rmtree(options.worker.spill_directory)
 
     def setUp(self):
-        self.worker_plasma_sock = '/tmp/plasma_%d_%d.sock' % (os.getpid(), id(Test))
         scheduler_port = str(get_next_port())
         proc_worker = subprocess.Popen([sys.executable, '-m', 'mars.worker',
                                         '-a', '127.0.0.1',
@@ -59,7 +58,6 @@ class Test(unittest.TestCase):
                                         '--level', 'debug',
                                         '--cache-mem', '16m',
                                         '--schedulers', '127.0.0.1:' + scheduler_port,
-                                        '--plasma-socket', self.worker_plasma_sock,
                                         '--ignore-avail-mem'])
         proc_scheduler = subprocess.Popen([sys.executable, '-m', 'mars.scheduler',
                                            '-H', '127.0.0.1',
@@ -114,8 +112,6 @@ class Test(unittest.TestCase):
         for p in procs:
             if p.poll() is None:
                 p.kill()
-
-        os.unlink(self.worker_plasma_sock)
 
         gevent.hub.Hub.NOT_ERROR = self.exceptions
