+++ mars/worker/tests/test_main.py
@@ -60,7 +60,6 @@ class Test(unittest.TestCase):
     def testExecuteWorker(self):
         import mars.tensor as mt
         mock_scheduler_addr = '127.0.0.1:%d' % get_next_port()
-        worker_plasma_sock = '/tmp/plasma_%d_%d.sock' % (os.getpid(), id(PromiseReplyTestActor))
         try:
 
             session_id = str(uuid.uuid4())
@@ -76,7 +75,6 @@ class Test(unittest.TestCase):
                                          '--schedulers', mock_scheduler_addr,
                                          '--cpu-procs', '1',
                                          '--cache-mem', '10m',
-                                         '--plasma-socket', worker_plasma_sock,
                                          '--ignore-avail-mem'])
                 worker_ips = []
 
@@ -127,5 +125,5 @@ class Test(unittest.TestCase):
                         break
                 if proc.poll() is None:
                     proc.kill()
-            if os.path.exists(worker_plasma_sock):
-                os.unlink(worker_plasma_sock)
+            if os.path.exists(options.worker.plasma_socket):
+                os.unlink(options.worker.plasma_socket)
